+++ packages/twenty-server/src/workspace/messaging/services/fetch-workspace-messages.module.ts
@@ -6,7 +6,7 @@ import { DataSourceModule } from 'src/metadata/data-source/data-source.module';
 import { MessagingModule } from 'src/workspace/messaging/messaging.module';
 import { MessagingProvidersModule } from 'src/workspace/messaging/providers/messaging-providers.module';
 import { FetchBatchMessagesService } from 'src/workspace/messaging/services/fetch-batch-messages.service';
-import { FetchWorkspaceMessagesService } from 'src/workspace/messaging/services/fetch-workspace-messages.service';
+import { GmailFullSync } from 'src/workspace/messaging/services/gmail-full-sync.service';
 import { RefreshAccessTokenService } from 'src/workspace/messaging/services/refresh-access-token.service';
 
 @Module({
@@ -18,10 +18,10 @@ import { RefreshAccessTokenService } from 'src/workspace/messaging/services/refr
     MessagingProvidersModule,
   ],
   providers: [
-    FetchWorkspaceMessagesService,
+    GmailFullSync,
     FetchBatchMessagesService,
     RefreshAccessTokenService,
   ],
-  exports: [FetchWorkspaceMessagesService, RefreshAccessTokenService],
+  exports: [GmailFullSync, RefreshAccessTokenService],
 })
 export class FetchWorkspaceMessagesModule {}
