+++ packages/twenty-server/src/workspace/messaging/services/utils.service.ts
@@ -1,118 +1,25 @@
 import { Injectable } from '@nestjs/common';
 
+import { EntityManager, DataSource } from 'typeorm';
 import { gmail_v1 } from 'googleapis';
 import { v4 } from 'uuid';
-import { DataSource, EntityManager } from 'typeorm';
 
 import { TypeORMService } from 'src/database/typeorm/typeorm.service';
 import { DataSourceService } from 'src/metadata/data-source/data-source.service';
-import { FetchBatchMessagesService } from 'src/workspace/messaging/services/fetch-batch-messages.service';
+import { DataSourceEntity } from 'src/metadata/data-source/data-source.entity';
 import {
   GmailMessage,
   Recipient,
 } from 'src/workspace/messaging/types/gmailMessage';
-import { MessageOrThreadQuery } from 'src/workspace/messaging/types/messageOrThreadQuery';
-import { DataSourceEntity } from 'src/metadata/data-source/data-source.entity';
-import { GmailClientProvider } from 'src/workspace/messaging/providers/gmail/gmail-client.provider';
 
 @Injectable()
-export class FetchWorkspaceMessagesService {
+export class Utils {
   constructor(
-    private readonly gmailClientProvider: GmailClientProvider,
     private readonly dataSourceService: DataSourceService,
     private readonly typeORMService: TypeORMService,
-    private readonly fetchBatchMessagesService: FetchBatchMessagesService,
   ) {}
 
-  public async fetchConnectedAccountThreads(
-    workspaceId: string,
-    connectedAccountId: string,
-    maxResults = 500,
-  ): Promise<void> {
-    const { workspaceDataSource, dataSourceMetadata, connectedAccount } =
-      await this.getDataSourceMetadataWorkspaceMetadataAndConnectedAccount(
-        workspaceId,
-        connectedAccountId,
-      );
-
-    const accessToken = connectedAccount.accessToken;
-    const refreshToken = connectedAccount.refreshToken;
-
-    if (!refreshToken) {
-      throw new Error('No refresh token found');
-    }
-
-    const gmailClient =
-      await this.gmailClientProvider.getGmailClient(refreshToken);
-
-    const threads = await gmailClient.users.threads.list({
-      userId: 'me',
-      maxResults,
-    });
-
-    const threadsData = threads.data.threads;
-
-    if (!threadsData || threadsData?.length === 0) {
-      return;
-    }
-
-    const { savedMessageIds, savedThreadIds } =
-      await this.getAllSavedMessagesIdsAndMessageThreadsIdsForConnectedAccount(
-        dataSourceMetadata,
-        workspaceDataSource,
-        connectedAccount.id,
-      );
-
-    const threadsToSave = threadsData.filter(
-      (thread) => thread.id && !savedThreadIds.includes(thread.id),
-    );
-
-    await this.saveMessageThreads(
-      threadsToSave,
-      dataSourceMetadata,
-      workspaceDataSource,
-      connectedAccount.id,
-    );
-
-    const threadQueries: MessageOrThreadQuery[] = threadsData.map((thread) => ({
-      uri: '/gmail/v1/users/me/threads/' + thread.id + '?format=minimal',
-    }));
-
-    const threadsWithMessageIds =
-      await this.fetchBatchMessagesService.fetchAllThreads(
-        threadQueries,
-        accessToken,
-      );
-
-    const messageIds = threadsWithMessageIds
-      .map((thread) => thread.messageIds)
-      .flat();
-
-    const messageIdsToSave = messageIds.filter(
-      (messageId) => !savedMessageIds.includes(messageId),
-    );
-
-    const messageQueries: MessageOrThreadQuery[] = messageIdsToSave.map(
-      (messageId) => ({
-        uri: '/gmail/v1/users/me/messages/' + messageId + '?format=RAW',
-      }),
-    );
-
-    const messagesResponse =
-      await this.fetchBatchMessagesService.fetchAllMessages(
-        messageQueries,
-        accessToken,
-      );
-
-    await this.saveMessages(
-      messagesResponse,
-      dataSourceMetadata,
-      workspaceDataSource,
-      connectedAccount,
-    );
-  }
-
-  private async saveMessageThreads(
+  public async saveMessageThreads(
     threads: gmail_v1.Schema$Thread[],
     dataSourceMetadata: DataSourceEntity,
     workspaceDataSource: DataSource,
@@ -135,7 +42,7 @@ export class FetchWorkspaceMessagesService {
     }
   }
 
-  private async saveMessages(
+  public async saveMessages(
     messages: GmailMessage[],
     dataSourceMetadata: DataSourceEntity,
     workspaceDataSource: DataSource,
@@ -219,7 +126,7 @@ export class FetchWorkspaceMessagesService {
     }
   }
 
-  async saveMessageRecipients(
+  public async saveMessageRecipients(
     recipients: Recipient[],
     dataSourceMetadata: DataSourceEntity,
     messageId: string,
@@ -258,7 +165,7 @@ export class FetchWorkspaceMessagesService {
     }
   }
 
-  private async getAllSavedMessagesIdsAndMessageThreadsIdsForConnectedAccount(
+  public async getAllSavedMessagesIdsAndMessageThreadsIdsForConnectedAccount(
     dataSourceMetadata: DataSourceEntity,
     workspaceDataSource: DataSource,
     connectedAccountId: string,
@@ -285,7 +192,7 @@ export class FetchWorkspaceMessagesService {
     };
   }
 
-  private async getDataSourceMetadataWorkspaceMetadataAndConnectedAccount(
+  public async getDataSourceMetadataWorkspaceMetadataAndConnectedAccount(
     workspaceId: string,
     connectedAccountId: string,
   ): Promise<{
